refactor(interactive-shell): remove the two utils/ wrapper packages
Neither surfaces/interactive_shell/utils/ nor
surfaces/interactive_shell/runtime/utils/ held anything general -- each
was a single-member directory whose name said only "things nobody
placed", inconsistent with the house style of role-named modules.

interactive_shell/utils/telemetry/ -> interactive_shell/telemetry/
interactive_shell/runtime/utils/input_policy.py ->
  interactive_shell/runtime/input_policy.py

Mirrored the move in tests/interactive_shell/ to match. No function
bodies changed -- import-path updates only, applied across the 14
production call sites, 21 test files, a comment in cli_parity.py, and
a path reference in AGENTS.md that all still spelled the old paths.

Verified: no utils/ dir remains under surfaces, no textual reference
to the old paths remains anywhere in the repo, a real interactive
shell session (via a pseudo-tty) starts, runs a genuine chat turn, and
exits cleanly with the turn correctly recorded in the prompt-log
JSONL, lint/format/typecheck clean, and the full
tests/interactive_shell/ + tests/core/agent/ suites pass
(1961 passed, 1 skipped).
