--- surfaces/interactive_shell/runtime/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Runtime utility helpers for interactive shell orchestration."""
