+++ surfaces/interactive_shell/telemetry/integration_snapshot.py
