+++ surfaces/interactive_shell/telemetry/investigation_analytics.py
