+++ surfaces/interactive_shell/telemetry/sinks/__init__.py
@@ -1,0 +1,8 @@
+"""Prompt logging sinks."""
+
+from surfaces.interactive_shell.telemetry.sinks.local_jsonl import (
+    append_prompt_log_record,
+)
+from surfaces.interactive_shell.telemetry.sinks.posthog_ai import capture_ai_generation
+
+__all__ = ["append_prompt_log_record", "capture_ai_generation"]
