--- surfaces/interactive_shell/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Shared interactive-shell utilities."""
