+++ tests/interactive_shell/telemetry/test_execution_integration.py
@@ -8,7 +8,7 @@ from surfaces.interactive_shell.runtime.core.turn_accounting import (
     ToolCallingTurnResult,
 )
 from surfaces.interactive_shell.session import Session
-from surfaces.interactive_shell.utils.telemetry import LlmRunInfo
+from surfaces.interactive_shell.telemetry import LlmRunInfo
 from tests.shared.harness_turn_driver import run_harness_turn
