+++ tests/interactive_shell/telemetry/test_posthog_ai.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from infrastructure.analytics.events import Event
-from surfaces.interactive_shell.utils.telemetry.sinks import posthog_ai
+from surfaces.interactive_shell.telemetry.sinks import posthog_ai
 
 
 def test_capture_ai_generation_uses_analytics_capture(monkeypatch) -> None:
