+++ tests/interactive_shell/test_terminal_runtime_turns.py
@@ -10,11 +10,11 @@ from rich.console import Console
 
 import surfaces.interactive_shell.runtime.slash_adapter as slash_adapter
 from core.llm.types import AgentLLMResponse, ToolCall
+from surfaces.interactive_shell.runtime import input_policy as loop_input_policy
 from surfaces.interactive_shell.runtime.core.turn_accounting import (
     ToolCallingTurnResult,
 )
 from surfaces.interactive_shell.runtime.turn_host import run_agent_turn_queue
-from surfaces.interactive_shell.runtime.utils import input_policy as loop_input_policy
 from surfaces.interactive_shell.session import Session
 from tests.core.agent.orchestration.action_execution_test_harness import (
     FakeActionLLM,
