+++ tests/interactive_shell/test_turn_outcome.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from surfaces.interactive_shell.utils.telemetry.turn_outcome import (
+from surfaces.interactive_shell.telemetry.turn_outcome import (
     format_investigation_outcome,
     format_investigation_terminal_outcome,
     format_terminal_turn_outcome,
